implemented rbac on imerc citl revision

+++ pages/api/imerc_citl_revision/[id]/index.tsx
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import iMERCCITLRevisionAbility from "@/services/ability/iMERCCITLRevisionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -20,7 +17,6 @@ export default async function handler(
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = iMERCCITLRevisionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -35,7 +31,6 @@ export default async function handler(
         {
           where: {
             AND: [
-              accessibleBy(ability).IMERCCITLRevision,
               {
                 id: {
                   equals: id,
@@ -63,12 +58,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "IMERCCITLRevision"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const iMERCCITLRevision = await prisma.iMERCCITLRevision.delete({
         where: {
